AutoGround
Made the ground auto fill under the tiles when using the auto ground prefabs.
Changed the graphics on the ship's chain. The scale of the game objects was 0.1 or 0.01 so when I put it back to 1 things stopped working. I fixed what I could but it still needs work. Sorry Forer.